+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 The file contains changelog for the project.
 
-## [0.0.2] - [Unreleased]
+## [0.1.1] - [Unreleased]
+
+- Added EmailLetter [CPB-22](https://github.com/DevIhor/Recruiter/pull/22)
+- Reset user's password [CPB-29](https://github.com/DevIhor/Recruiter/pull/27)
+- Added Vacancy CRUD api endpoint [CPB-33](https://github.com/DevIhor/Recruiter/pull/26)
+- Created events views, factories and tests [CPB-35](https://github.com/DevIhor/Recruiter/pull/24)
+- Merged settings into a single file [CPB-37](https://github.com/DevIhor/Recruiter/pull/23)
+- Updated readme [CPB-45](https://github.com/DevIhor/Recruiter/pull/25)
+- Updated changelog [CPB-46]
+
+## [0.1.0] - [Unreleased] - 07.09.2022
 
 - Project skeleton was created [CPB-1](https://github.com/DevIhor/Recruiter/pull/1)
 - Updated Docs [CPB-2](https://github.com/DevIhor/Recruiter/pull/9)
@@ -15,6 +25,12 @@ The file contains changelog for the project.
 - Environment variables were submitted [CPB-9](https://github.com/DevIhor/Recruiter/pull/1)
 - Added Static and Media Configurations [CPB-12](https://github.com/DevIhor/Recruiter/pull/6)
 - Candidate model was created [CPB-14](https://github.com/DevIhor/Recruiter/pull/5)
+- Created recruiter and reviewer groups [CPB-16](https://github.com/DevIhor/Recruiter/pull/20)
 - Vacancy model was created [CPB-17](https://github.com/DevIhor/Recruiter/pull/10)
+- Candidate-Vacancy relation was added [CPB-18](https://github.com/DevIhor/Recruiter/pull/17)
+- Added Event model, added EventType model [CPB-19](https://github.com/DevIhor/Recruiter/pull/14)
 - EmailTemplate model was created [CPB-21](https://github.com/DevIhor/Recruiter/pull/12)
 - Setuped DRF and JWT [CPB-26](https://github.com/DevIhor/Recruiter/pull/13)
+- Sign In API was created [CPB-27](https://github.com/DevIhor/Recruiter/pull/19)
+- Sign Up API was created [CPB-28](https://github.com/DevIhor/Recruiter/pull/18)
+- Divide requirements into dev and prod [CPB-36](https://github.com/DevIhor/Recruiter/pull/16)